Move isRecordableResult function into root lib folder
