+++ packages/core/lib/isRecordableResult.test.ts
@@ -1,4 +1,4 @@
-import type { Result } from "../../types/AnnotatedHearingOutcome"
+import type { Result } from "../types/AnnotatedHearingOutcome"
 
 import isRecordableResult from "./isRecordableResult"
