+++ packages/core/lib/isRecordableResult.ts
@@ -1,6 +1,6 @@
-import type { Result } from "../../types/AnnotatedHearingOutcome"
+import type { Result } from "../types/AnnotatedHearingOutcome"
 
-import nonRecordableResultCodes from "../../lib/nonRecordableResultCodes"
+import nonRecordableResultCodes from "./nonRecordableResultCodes"
 
 const isRecordableResult = (result: Result): boolean =>
   !!result.PNCDisposalType && !nonRecordableResultCodes.includes(result.PNCDisposalType)
