+++ packages/core/phase2/lib/areResultsMatchingAPncDisposal.ts
@@ -2,7 +2,7 @@ import type { Offence } from "../../types/AnnotatedHearingOutcome"
 import type { PncDisposal } from "../../types/PncQueryResult"
 import type { CheckExceptionFn } from "./areResultsMatchingPncAdjudicationAndDisposals"
 
-import isRecordableResult from "./isRecordableResult"
+import isRecordableResult from "../../lib/isRecordableResult"
 import isResultMatchingAPncDisposal from "./isResultMatchingAPncDisposal"
 
 const areResultsMatchingAPncDisposal = (
